+++ stts/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.9.2
 * fix(Windows): Throw `PlatformException` on error.
+fix(Android/STT): Code cleanup with some improvement.
 
 ## 0.9.1
 * feat(Android/STT): Add `muteSystemSounds` to mute default system beep sounds.