+++ stts/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.9.2
 * fix(Windows): Throw `PlatformException` on error.
-fix(Android/STT): Code cleanup with some improvement.
+* fix(Android/STT): Code cleanup with some improvement.
+* fix(iOS/macOS): Small code fixes/improvements.
 
 ## 0.9.1
 * feat(Android/STT): Add `muteSystemSounds` to mute default system beep sounds.